# Conflicts:
#	prehension/pressure_sensors/preprocess_pressure_sensors.py

+++ prehension/visualization/session_data_visualization.py
@@ -27,6 +27,8 @@ import shutil
 import glob
 import pandas as pd
 import matplotlib.colors as mcolors
+import matplotlib
+matplotlib.use('Agg')  # Use non-interactive backend
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -438,12 +440,21 @@ class SessionWrapper:
             # defined vars:
             # [dst_session_dir] - raw_(training)_server_session: the expected directory name after
             # transfer
+
+            # Check that we don't somehow transfer to a raw from a processed dir or vice-versa
+            if 'raw' in src_session_dir.lower() and 'proc' in dst_parent_dir.lower():
+                raise ValueError('Src path contains the string raw, but found proc in'
+                                    'destination directory. Ensure the preset values are correct')
+            if 'proc' in src_session_dir.lower() and 'raw' in dst_parent_dir.lower():
+                raise ValueError('Src path contains the string proce, but found raw in'
+                                    'destination directory. Ensure the preset values are correct')
+
             dst_session_dir = os.path.join(dst_parent_dir, os.path.basename(src_session_dir))
 
             # logic:
             src_exists = os.path.exists(src_session_dir)
             if not src_exists:  # This should always exist
-                raise Exception(f"Source session directory {src_session_dir} does not exist")
+                raise FileNotFoundError(f"Source session directory {src_session_dir} does not exist")
 
             # Check if dst_session_dir exists
             already_transferred = False
@@ -452,17 +463,26 @@ class SessionWrapper:
                 if src_contents.issubset(dst_contents):
                     already_transferred = True
 
+            if already_transferred:
+                rs('Folder already transfered')
+                return
+
             # To keep track of transfer status
             transferred = False
 
             if not already_transferred or overwrite:
                 # rs('transferring')
                 # copy src to dst_parent_dir but remove dst first
+
+
                 if os.path.exists(dst_session_dir):
-                    try:
-                        shutil.rmtree(dst_session_dir, ignore_errors=False)
-                    except:
-                        ws(f"failed to remove {dst_session_dir}")
+                    ws('Warning, dst session directory already exists, overwriting...')
+
+                # if os.path.exists(dst_session_dir):
+                #     try:
+                #         shutil.rmtree(dst_session_dir, ignore_errors=False)
+                #     except:
+                #         ws(f"failed to remove {dst_session_dir}")
 
                 try:
                     shutil.move(src_session_dir, dst_parent_dir)
@@ -476,13 +496,13 @@ class SessionWrapper:
                 src_contents, dst_contents = _folder_contents_set(src_session_dir, dst_session_dir)
                 assert src_contents.issubset(dst_contents)
 
-            # now remove src locally because we should have already transferred
-            if os.path.exists(src_session_dir):
-                assert os.path.exists(dst_session_dir), "dir not found: {dst_session_dir}"
-                try:
-                    shutil.rmtree(src_session_dir, ignore_errors=False)
-                except:
-                    ws(f"failed to remove {src_session_dir}")
+                # now remove src locally because we should have already transferred
+                if os.path.exists(src_session_dir):
+                    assert os.path.exists(dst_session_dir), "dir not found: {dst_session_dir}"
+                    try:
+                        shutil.rmtree(src_session_dir, ignore_errors=False)
+                    except:
+                        ws(f"failed to remove {src_session_dir}")
 
         # Transfer from experiment to training server
         _move_helper(self.raw_ss, raw_training_server)
@@ -493,6 +513,8 @@ class SessionWrapper:
         if not self.has_meta:
             return False
         # Check if it is an experiment or not
+        # Consider it an experiment if it has videos directory and raw_ps directory
+        # Neural is optional right now
         experiment_expected_dirnames = [os.path.join(self.raw_ss, self.mstruct["videos_dir"]),
                                         os.path.join(self.raw_ss, self.mstruct["raw_ps_dir"])]
         is_training_session = not all([os.path.exists(dname) for dname
@@ -539,7 +561,10 @@ class SessionWrapper:
             # Extract the set number if available, otherwise return -1
             set_number = int(match.group(2)) if match.group(2) else -1
 
-        return current_date, set_number
+            return current_date, set_number
+
+        else:
+            raise ValueError(f'Could not parse date from folder name: {folder}')
 
     def plot_force_trace(self):
         tp_path = self.mstruct['timepoint_csv_filename']
@@ -568,6 +593,7 @@ class SessionWrapper:
 
         # Return and THEN bind results to trial object
         return (times, forces_summed)
+
 
     def plot_force_traces(
             self,
